[UI-12] Rename `Element` and `Panel` in `Sidebar` to `Sidebar*`
