+++ packages/sidebar/src/SidebarElement.js
