+++ packages/sidebar/src/SidebarPanel.js
